GEOS-3830) UpdateSequence is not getting updated
git-svn-id: http://svn.codehaus.org/geoserver/trunk@14062 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 8dec09444a47c5bf0318f638e76ec4bfd5c1da69 [formerly 5182643de2f4b20d258561c5554a37c3b2b51dfb]
Former-commit-id: 0816ca9a162694e9bcad1409d23ab65faf30294f

+++ src/wcs1_0/src/main/java/org/geoserver/wcs/response/Wcs10CapsTransformer.java
@@ -122,11 +122,11 @@ public class Wcs10CapsTransformer extends TransformerBase {
             this.request = (GetCapabilitiesType) o;
 
             // check the update sequence
-            final int updateSequence = wcs.getGeoServer().getGlobal().getUpdateSequence();
-            int requestedUpdateSequence = -1;
+            final long updateSequence = wcs.getGeoServer().getGlobal().getUpdateSequence();
+            long requestedUpdateSequence = -1;
             if (request.getUpdateSequence() != null) {
                 try {
-                    requestedUpdateSequence = Integer.parseInt(request.getUpdateSequence());
+                    requestedUpdateSequence = Long.parseLong(request.getUpdateSequence());
                 } catch (NumberFormatException e) {
                     if (request.getUpdateSequence().length() == 0)
                         requestedUpdateSequence = 0;
